backend/requirements.txt

fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-anthropic>=0.1.4
langchain-openai>=0.0.5
langgraph>=0.1.0  # Upgrade to 1.0+ compatible version
mcp>=1.0.0 # Add MCP SDK
supabase>=2.3.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.27.0
aiofiles>=23.2.0
python-multipart>=0.0.9
google-generativeai>=0.4.0
pandas>=2.1.4
openpyxl>=3.1.2
unidecode>=1.3.7
openai>=1.12.0
pdfplumber>=0.10.3
pypdf>=3.17.4
pytesseract>=0.3.10
pytest>=8.0.0
pytest-asyncio>=0.23.0
slowapi>=0.1.9
paper-search-mcp>=0.1.0
